+++ liblzr/src/interpolator.hpp
@@ -8,10 +8,10 @@ namespace lzr {
 
 //generates extra points on a line
 //only adds the interior points, not the endpoints
-int interp_line(Frame& working, double max_distance, interpolation_func func, Point start, Point end);
+int interp_line(Frame& working, float max_distance, interpolation_func func, Point start, Point end);
 
 //interpolates only lit lines. Uses the function above
-int interpolate(Frame& frame, double max_distance, interpolation_func func);
+int interpolate(Frame& frame, float max_distance, interpolation_func func);
 
 } // namespace lzr
